diagnosis 컨트롤러
- 학생 목록 보기
- 학생 상세 보기
- 학습지 목록 보기
- 학습지 상세 보기 : 학습지 id에 따른 문항 이미지 보기 (테이블 조인)

+++ diagnosis/src/main/java/com/mmt/diagnosis/domain/Test.java
@@ -11,10 +11,10 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 public class Test {
 
-    private int testId;
+    private Long testId;
     private String testName;
     private String testComments;
     private LocalDateTime testTimestamp;
-    private int diagnosticTestId;
+    private Long diagnosticTestId;
 
 }
