+++ diagnosis/src/main/java/com/mmt/diagnosis/dto/test/TestResponse.java
@@ -7,10 +7,10 @@ import java.time.LocalDateTime;
 @Data
 public class TestResponse {
 
-    private int testId;
+    private Long testId;
     private String testName;
     private String testComments;
     private LocalDateTime testTimestamp;
-    private int diagnosticTestId;
+    private Long diagnosticTestId;
 
 }
